+++ keyers.h
@@ -14,6 +14,7 @@ public:
     virtual void EndTx();
     virtual void BeginTx(int relay);
     virtual void EndTx(int relay);
+    
 };
 
 class Keyer {
@@ -26,7 +27,7 @@ public:
     virtual bool TxClosed(int relay);
     virtual void Tx(int relay, bool closed);
     virtual void Key(Paddle key, bool pressed);
-    virtual void Tick(unsigned int millis);
+    virtual void Tick(unsigned long millis);
 };
 
 Keyer *GetKeyerByNumber(int n, Transmitter *output);
